Refactor Facility Groupings to be in its own file
