+++ docs/Facilities.md
@@ -8,9 +8,7 @@ of, information about different facilities
 
 * [Facilities](#Facilities)
     * [new Facilities(sdk, request)](#new_Facilities_new)
-    * [.addFacilityToGrouping(facilityGroupingId, facilityId)](#Facilities+addFacilityToGrouping) ⇒ <code>Promise</code>
     * [.create(facility)](#Facilities+create) ⇒ <code>Promise</code>
-    * [.createGrouping(facilityGrouping)](#Facilities+createGrouping) ⇒ <code>Promise</code>
     * [.createOrUpdateInfo(facilityId, update)](#Facilities+createOrUpdateInfo) ⇒ <code>Promise</code>
     * [.delete(facilityId)](#Facilities+delete) ⇒ <code>Promise</code>
     * [.get(facilityId)](#Facilities+get) ⇒ <code>Promise</code>
@@ -27,29 +25,6 @@ of, information about different facilities
 | sdk | <code>Object</code> | An instance of the SDK so the module can communicate with other modules |
 | request | <code>Object</code> | An instance of the request module tied to this module's audience. |
 
-<a name="Facilities+addFacilityToGrouping"></a>
-
-### contxtSdk.facilities.addFacilityToGrouping(facilityGroupingId, facilityId) ⇒ <code>Promise</code>
-Adds a facility to a facility grouping
-
-API Endpoint: '/groupings/:facilityGroupingId/facilities/:facilityId'
-Method: POST
-
-**Kind**: instance method of [<code>Facilities</code>](#Facilities)  
-**Fulfill**: [<code>FacilityGroupingFacility</code>](./Typedefs.md#FacilityGroupingFacility) Information about the new facility/grouping relationship  
-**Reject**: <code>Error</code>  
-
-| Param | Type | Description |
-| --- | --- | --- |
-| facilityGroupingId | <code>string</code> | UUID corresponding with a facility grouping |
-| facilityId | <code>number</code> |  |
-
-**Example**  
-```js
-contxtSdk.facilities.addFacilityToGrouping('b3dbaae3-25dd-475b-80dc-66296630a8d0', 4)
-  .then((grouping) => console.log(grouping));
-  .catch((err) => console.log(err));
-```
 <a name="Facilities+create"></a>
 
 ### contxtSdk.facilities.create(facility) ⇒ <code>Promise</code>
@@ -85,40 +60,6 @@ contxtSdk.facilities
     organizationId: 25
   })
   .then((facilities) => console.log(facilities));
-  .catch((err) => console.log(err));
-```
-<a name="Facilities+createGrouping"></a>
-
-### contxtSdk.facilities.createGrouping(facilityGrouping) ⇒ <code>Promise</code>
-Creates a new facility grouping
-
-API Endpoint: '/groupings'
-Method: POST
-
-**Kind**: instance method of [<code>Facilities</code>](#Facilities)  
-**Fulfill**: [<code>FacilityGrouping</code>](./Typedefs.md#FacilityGrouping) Information about the new facility grouping  
-**Reject**: <code>Error</code>  
-
-| Param | Type | Default | Description |
-| --- | --- | --- | --- |
-| facilityGrouping | <code>Object</code> |  |  |
-| [facilityGrouping.description] | <code>string</code> |  |  |
-| [facilityGrouping.isPrivate] | <code>boolean</code> | <code>false</code> |  |
-| facilityGrouping.name | <code>string</code> |  |  |
-| facilityGrouping.organizationId | <code>string</code> |  | UUID |
-| [facilityGrouping.parentGroupingId] | <code>string</code> |  | UUID |
-
-**Example**  
-```js
-contxtSdk.facilities
-  .createGrouping({
-    description: 'US States of CT, MA, ME, NH, RI, VT',
-    isPrivate: false,
-    name: 'New England, USA',
-    organization_id: '61f5fe1d-d202-4ae7-af76-8f37f5bbeec5'
-    parent_grouping_id: 'e9f8f89c-609c-4c83-8ebc-cea928af661e'
-  })
-  .then((grouping) => console.log(grouping));
   .catch((err) => console.log(err));
 ```
 <a name="Facilities+createOrUpdateInfo"></a>
